+++ OneBlink.SDK/models/EndpointConfiguration.cs
@@ -1,0 +1,26 @@
+namespace OneBlink.SDK.Model
+{
+    public class EndpointConfiguration
+    {
+        public string apiId
+        {
+            get; set;
+        }
+        public string apiEnvironment
+        {
+            get; set;
+        }
+        public string apiEnvironmentRoute
+        {
+            get; set;
+        }
+        public string url
+        {
+            get; set;
+        }
+        public string secret
+        {
+            get; set;
+        }
+    }
+}