+++ OneBlink.SDK/models/FormExternalIdGenerationConfiguration.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+
+namespace OneBlink.SDK.Model
+{
+    public class FormExternalIdGenerationConfiguration : EndpointConfiguration
+    {
+        public List<FormExternalIdGenerationReceiptComponent> receiptComponents
+        {
+            get; set;
+        }
+    }
+
+}